+++ src/core/rocketchat-servers.h
@@ -18,7 +18,6 @@ struct _ROCKETCHAT_SERVER_REC {
 	GQueue *message_queue;
 	GString *buffer;
 	GHashTable *result_callbacks;
-	GHashTable *sent_messages;
 	GHashTable *rooms;
 	char *userId;
 	char *tmid;